pyproject.toml
[tool.poetry]
name = "rlberry"
version = "v0.7.3.post2.dev0+5d8dce8 "
description = "An easy-to-use reinforcement learning library for research and education"
authors = ["Omar Darwiche Domingues, Yannis Flet-Berliac, Edouard Leurent, Pierre Menard, Xuedong Shang"]
homepage = "https://github.com/rlberry-py"
repository = "https://github.com/rlberry-py"
readme = "README.md"
license="LICENSE"
classifiers=[
        "Development Status :: 4 - Beta",
        "Intended Audience :: Science/Research",
        "Intended Audience :: Education",
        "Topic :: Scientific/Engineering :: Artificial Intelligence",
        "Programming Language :: Python :: 3.11",
        "Programming Language :: Python :: 3 :: Only",
        "License :: OSI Approved :: MIT License",
        "Operating System :: OS Independent",
    ]

[tool.poetry.dependencies]
python = "^3.9, <3.13"
scipy = "*"
pygame-ce = "*"
matplotlib = "*"
gymnasium = {version="^0.29.1",  extras=["atari", "accept-rom-license"]}
dill = "*"
docopt = "*"
pyyaml = "*"
tqdm = "*"
adastop = "*"
moviepy = "*"
pyopengl = "*"
optuna ={version="*", optional=true}
ffmpeg-python = {version="*", optional=true}
opencv-python = {version="*", optional=true}
ale-py = {version="*", optional=true}
stable-baselines3 = {version="*", optional=true}
tensorboard = {version="*", optional=true}
torch = {version=">=2.0.0, !=2.0.1, !=2.1.0", optional=true}
pandas = "*"
scikit-fda = {version="^0.9", optional=true}
sphinx = {version="6.2.1", optional=true}
sphinx-gallery = { version= "^0.14.0", optional=true}
sphinx-math-dollar = {version="^1.2.1", optional=true}
sphinxcontrib-video = {version="0.2.0", optional=true}
sphinx-copybutton = "*"
numpydoc = {version="^1.6.0", optional=true}
myst-parser = {version="^2.0.0", optional=true}


########### hack : some dependencies are not compatible #############
# dependency of scikit-fda :
fdasrsf = {version="2.5.2", optional=true}   # new version (2.5.8) : crash all our CI  (21-02-2024)
multimethod = {version="1.10", optional=true}     # new version 1.11 : crash our tests in "rlberry/manager/tests/test_plot.py"  (21-02-2024)
###############################################################


[tool.poetry.extras]
torch = ["opencv-python", "ale-py", "stable-baselines3", "tensorboard", "torch"]
extras = ["optuna", "ffmpeg-python", "scikit-fda"]
doc = ["sphinx", "sphinx-gallery", "sphinx-copybutton", "sphinx-math-dollar", "numpydoc", "myst-parser", "sphinxcontrib-video", "matplotlib"]

[tool.poetry.group.dev]
optional = true

[tool.poetry.group.dev.dependencies]
pytest = "^7.4.3"
pytest-xprocess = "^0.23.0"
codecov = "^2.1.13"
black = "23.9.1"
pre-commit = "^3.5.0"
rlberry-research = {git = "https://github.com/rlberry-py/rlberry-research.git"}
rlberry-scool = {git = "https://github.com/rlberry-py/rlberry-scool.git"}
sphinxcontrib-video = {version="0.2.0", optional=true}
sphinx = {version="6.2.1", optional=true}
sphinx-gallery = { version= "^0.14.0", optional=true}
sphinx-math-dollar = {version="^1.2.1", optional=true}
numpydoc = {version="^1.6.0", optional=true}
myst-parser = {version="^2.0.0", optional=true}
sphinx-copybutton = "*"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
